+++ src/main/steps/applicant1/hub-page/sole/content.ts
@@ -6,7 +6,7 @@ import { TranslationFn } from '../../../../app/controller/GetController';
 import type { CommonContent } from '../../../common/common.content';
 import { FINALISING_YOUR_APPLICATION, HOW_YOU_CAN_PROCEED } from '../../../urls';
 
-import { getSoleHubTemplate } from './soleTemplateChooser';
+import { getSoleHubTemplate } from './soleTemplateSelector';
 
 const en = ({ isDivorce, partner, userCase }: CommonContent, alternativeServiceType: AlternativeServiceType) => ({
   aosAwaitingOrDrafted: {
